% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cointoss_methods.R
\name{plot.cointoss}
\alias{plot.cointoss}
\title{Plot method for cointoss object}
\usage{
\method{plot}{cointoss}(x, type = c("bar", "path"), ...)
}
\arguments{
\item{x}{A cointoss object}

\item{type}{Type of plot: "bar" (default) or "path"}

\item{...}{Additional arguments passed to plotting functions}
}
\description{
Plot method for cointoss object
}
