% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cointoss_methods.R
\name{print.cointoss}
\alias{print.cointoss}
\title{Print method for cointoss object}
\usage{
\method{print}{cointoss}(x, ...)
}
\arguments{
\item{x}{A cointoss object}

\item{...}{Ignored}
}
\description{
Print method for cointoss object
}
