% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cointoss_methods.R
\name{summary.cointoss}
\alias{summary.cointoss}
\title{Summary method for cointoss object}
\usage{
\method{summary}{cointoss}(object, ...)
}
\arguments{
\item{object}{A cointoss object}

\item{...}{Ignored}
}
\value{
A named list of proportions
}
\description{
Summary method for cointoss object
}
