% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coin_toss.R
\name{toss_coin}
\alias{toss_coin}
\title{Simulate coin tosses}
\usage{
toss_coin(n = 10)
}
\arguments{
\item{n}{Number of tosses to simulate}
}
\value{
An object of class \code{cointoss}
}
\description{
Simulate coin tosses
}
\examples{
toss_coin(10)
}
